backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Datenbank
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Sicherheit & Verschlüsselung
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyotp==2.9.0  # MFA/TOTP

# JWT & Auth
pyjwt==2.8.0
python-dotenv==1.0.0

# Validierung
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
structlog==24.1.0

# CSV-Import (SimpliMed)
pandas==2.1.4
openpyxl==3.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Redis (Session-Management)
redis==5.0.1
hiredis==2.3.2

# CORS
python-cors==1.0.0
